requirements.txt
absl-py==2.3.0
cycler==0.12.1
fonttools==4.58.2
gast==0.4.0
h5py==3.14.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==1.23.5
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==3.19.6
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
six==1.17.0
tensorflow-gpu==2.10.0
typing_extensions==4.14.0
wrapt==1.17.2
zipp==3.23.0
